requirements.txt
py-cord==2.6.1
python-dotenv==1.2.1